Create OLED_firmware_test_I_modified_success.ino
This is the most current version of the OLED subsystem code, as of March 20th. It uses the group authentication key and is on the same database as other components. It is not a final product because this modifies virtual pin 0 (lock and unlock state) and virtual pin 2 (status of lock and unlock state). In the future this should only read those values, not write them to the databse. Super chart is working for "live" data range.